KInda Improved It

+++ src/net/psj/Particles/ParticleClone.java
@@ -20,8 +20,8 @@ public class ParticleClone extends Particle{
 		boolean ret = super.update();
 		if(ctype!=-0)
 		{
-			for(int x2 = x-1; x2 < x+1; x2++)
-				for(int y2 = y-1; y2 < y+1; y2++)
+			for(int x2 = (int)x-1; x2 < x+1; x2++)
+				for(int y2 = (int)y-1; y2 < y+1; y2++)
 					PowderSimJ.ptypes.create_part(x2,y2,ctype,false);
 		}
 		else
@@ -31,7 +31,7 @@ public class ParticleClone extends Particle{
 				for (ry=-1; ry<2; ry++)
 					if (x+rx>=0 && y+ry>0 && x+rx<PowderSimJ.width && y+ry<PowderSimJ.height && (rx!=0 || ry!=0))
 					{
-						Particle r = ParticleData.parts[ParticleData.pmap[y+ry][x+rx]];
+						Particle r = ParticleData.parts[ParticleData.pmap[(int)y+ry][(int)x+rx]];
 						if (r==null)
 							continue;
 						if(r.id==0)
